+++ src/api/namespaces/atomicmarket/index.ts
@@ -17,6 +17,7 @@ import { statsEndpoints } from './routes/stats';
 import ApiNotificationReceiver from '../../notification';
 import { buyoffersEndpoints, buyofferSockets } from './routes/buyoffers';
 import { assetsEndpoints } from './routes/assets';
+import { ActionHandlerOptions } from '../../actionhandler';
 
 export type AtomicMarketNamespaceArgs = {
     atomicmarket_account: string
@@ -54,6 +55,8 @@ export enum BuyofferApiState {
     ACCEPTED = 3,
     INVALID = 4
 }
+
+export type AtomicMarketActionHandlerOptions = ActionHandlerOptions<AtomicMarketNamespace>;
 
 export class AtomicMarketNamespace extends ApiNamespace {
     static namespaceName = 'atomicmarket';
