+++ src/api/namespaces/atomicmarket/routes/handlers/sales.ts
@@ -2,9 +2,9 @@ import { RequestParams } from '../../../utils';
 import { fillSales } from '../../filler';
 import { formatSale } from '../../format';
 import { ApiError } from '../../../../error';
-import { HandlerOptions } from '../wrapper';
+import { AtomicMarketActionHandlerOptions } from '../../index';
 
-export async function getSaleAction(params: RequestParams, options: HandlerOptions): Promise<any> {
+export async function getSaleAction(params: RequestParams, options: AtomicMarketActionHandlerOptions): Promise<any> {
     const query = await options.db.query(
         'SELECT * FROM atomicmarket_sales_master WHERE market_contract = $1 AND sale_id = $2',
         [options.core.args.atomicmarket_account, params.sale_id]
