+++ src/api/server.ts
@@ -16,6 +16,9 @@ import { expressRedisCache, ExpressRedisCacheHandler } from '../utils/cache';
 import { eosioTimestampToDate } from '../utils/eosio';
 import * as swagger from 'swagger-ui-express';
 import { getOpenApiDescription, LogSchema } from './docs';
+import { respondApiError } from './utils';
+import { ActionHandler } from './actionhandler';
+import { ApiNamespace } from './namespaces/interfaces';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const packageJson: any = require('../../package.json');
@@ -262,6 +265,36 @@ export class WebServer {
 
         this.express.use(router);
     }
+
+    wrapJSONHandler(handler: ActionHandler, core: ApiNamespace): express.Handler {
+        const server = this.server;
+
+        function mergeRequestData(req: express.Request): {[key: string]: any} {
+            return Object.assign({}, req.query || {}, req.body || {}, req.params || {});
+        }
+
+        return async (req: express.Request, res: express.Response): Promise<void> => {
+            try {
+                const params = mergeRequestData(req);
+
+                const options = {
+                    db: {
+                        async query<T = any>(queryText: string, values?: any[]): Promise<QueryResult<T>> {
+                            return server.query(queryText, values);
+                        },
+                    },
+                    core,
+                };
+
+                const result = await handler(params, options);
+
+                res.json({success: true, data: result, query_time: Date.now()});
+            } catch (error) {
+                respondApiError(res, error);
+            }
+        };
+    }
+
 }
 
 export class SocketServer {
